+++ crypto/services/src/main/java/org/hyperledger/besu/cryptoservices/pluginadapter/SecurityModuleServiceImpl.java
@@ -12,7 +12,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.services;
+package org.hyperledger.besu.cryptoservices.pluginadapter;
 
 import org.hyperledger.besu.plugin.services.SecurityModuleService;
 import org.hyperledger.besu.plugin.services.securitymodule.SecurityModule;
