+++ settings.gradle
@@ -36,6 +36,7 @@ include 'consensus:qbft'
 include 'consensus:qbft-core'
 include 'datatypes'
 include 'crypto:algorithms'
+include 'crypto:plugin-api'
 include 'crypto:services'
 include 'datatypes'
 include 'ethereum:api'
